+++ HISTORY.rst
@@ -2,6 +2,9 @@
 History
 =======
 
+* Thu Oct 7 2021 Jelmer Veenstra <jelmer.veenstra@deltares.nl> 2.2.89
+- generalized imports, now only 'import hatyan' necessary, all modules can be loaded from there
+
 * Mon Oct 4 2021 Jelmer Veenstra <jelmer.veenstra@deltares.nl> 2.2.88
 - bugfix for foreman frequencies, were actually retrieved from schureman